pyproject.toml
[project]
name = "ercot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.28.1",
    "onepassword-sdk>=0.2.0",
    "polars>=1.24.0",
    "requests>=2.32.3",
]
